move ClientProperties to module service-client
