+++ mqtt/client/build.gradle.kts
@@ -8,7 +8,6 @@ kotlin {
     sourceSets {
         commonMain {
             dependencies {
-//                ver my
                 api(project(":mqtt:codec"))
 
                 implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:${ver.jetbrains.coroutines}")
