+++ service-client/src/main/aidl/com/zibi/service/client/IServiceBroker.aidl
@@ -2,7 +2,7 @@ interface IServiceBroker {
     /**
     *   control threads
     */
-      void startSearchKey();
-      void stopSearchKey(); //back how many key check
+      void startMqttClient();
+      void stopMqttClient();
       boolean startStop();
 }
